Do hometask

+++ src/services/contacts.js
@@ -8,11 +8,12 @@ export const getAllContacts = async ({
                                        sortOrder = SORT_ORDER.ASC,
                                        sortBy = '_id',
                                        filter = {},
+                                       userId
                                      }) => {
   const limit = perPage;
   const skip = (page - 1) * perPage;
 
-  const contactsQuery = ContactsCollection.find();
+  const contactsQuery = ContactsCollection.find({ userId });
 
   if (filter.isFavourite !== undefined) {
     contactsQuery.where('isFavourite').equals(filter.isFavourite);
@@ -41,23 +42,21 @@ export const getAllContacts = async ({
   };
 };
 
-export const getContactById = async (contactId) => {
-  return ContactsCollection.findById(contactId);
+export const getContactById = async (contactId, userId) => {
+  return ContactsCollection.findOne({ _id: contactId, userId });
 };
 
 export const createContact = async (payload) => {
   return ContactsCollection.create(payload);
 };
 
-export const deleteContact = async (contactId) => {
-  return ContactsCollection.findOneAndDelete({
-    _id: contactId,
-  });
+export const deleteContact = async (contactId, userId) => {
+  return ContactsCollection.findOneAndDelete({ _id: contactId, userId });
 };
 
 export const updateContact = async (contactId, payload, options = {}) => {
   const rawResult = await ContactsCollection.findOneAndUpdate(
-    { _id: contactId },
+    { _id: contactId, userId: options.userId },
     payload,
     {
       new: true,
